+++ clr/DNP3CLRAdapter/MasterDataObserverAdapter.h
@@ -18,11 +18,11 @@ public:
 
 	MasterDataObserverAdapter(DNP3::Interface::IDataObserver ^ proxy);
 
-	void Update(const opendnp3::Binary& arPoint, size_t aIndex);
-	void Update(const opendnp3::Analog& arPoint, size_t aIndex);
-	void Update(const opendnp3::Counter& arPoint, size_t aIndex);
-	void Update(const opendnp3::ControlStatus& arPoint, size_t aIndex);
-	void Update(const opendnp3::SetpointStatus& arPoint, size_t aIndex);
+	void Update(const opendnp3::Binary& arPoint, uint16_t aIndex);
+	void Update(const opendnp3::Analog& arPoint, uint16_t aIndex);
+	void Update(const opendnp3::Counter& arPoint, uint16_t aIndex);
+	void Update(const opendnp3::ControlStatus& arPoint, uint16_t aIndex);
+	void Update(const opendnp3::SetpointStatus& arPoint, uint16_t aIndex);
 
 private:
