+++ cpp/opendnp3tests/src/ComparingDataObserver.cpp
@@ -91,27 +91,27 @@ void ComparingDataObserver::End()
 
 }
 
-void ComparingDataObserver::Update(const Binary& arPoint, size_t aIndex)
+void ComparingDataObserver::Update(const Binary& arPoint, uint16_t aIndex)
 {
 	this->UpdateAny<Binary>(arPoint, aIndex, mpObserver->mBinaryMap, mBinaryMap);
 }
 
-void ComparingDataObserver::Update(const Analog& arPoint, size_t aIndex)
+void ComparingDataObserver::Update(const Analog& arPoint, uint16_t aIndex)
 {
 	this->UpdateAny<Analog>(arPoint, aIndex, mpObserver->mAnalogMap, mAnalogMap);
 }
 
-void ComparingDataObserver::Update(const Counter& arPoint, size_t aIndex)
+void ComparingDataObserver::Update(const Counter& arPoint, uint16_t aIndex)
 {
 	this->UpdateAny<Counter>(arPoint, aIndex, mpObserver->mCounterMap, mCounterMap);
 }
 
-void ComparingDataObserver::Update(const ControlStatus& arPoint, size_t aIndex)
+void ComparingDataObserver::Update(const ControlStatus& arPoint, uint16_t aIndex)
 {
 	this->UpdateAny<ControlStatus>(arPoint, aIndex, mpObserver->mControlStatusMap, mControlStatusMap);
 }
 
-void ComparingDataObserver::Update(const SetpointStatus& arPoint, size_t aIndex)
+void ComparingDataObserver::Update(const SetpointStatus& arPoint, uint16_t aIndex)
 {
 	this->UpdateAny<SetpointStatus>(arPoint, aIndex, mpObserver->mSetpointStatusMap, mSetpointStatusMap);
 }
